feat: 검색페이지 Hydration 기능 추가
